Factor out AlterRoleTask from TransportAlterRoleAction
